Fixed [SF bug 1594664] PointF X and Y values were being cast to an int.  Removed the cast
Full patch submitted via email by Pawel Konieczny
Signed off by: jamcquay

+++ src/Axis.cs
@@ -844,7 +844,7 @@ namespace NPlot
 				(float)(prop * (_physicalMax.X - _physicalMin.X)),
 				(float)(prop * (_physicalMax.Y - _physicalMin.Y)) );
 
-			return new PointF( (int)(_physicalMin.X + offset.X), (int)(_physicalMin.Y + offset.Y) );
+			return new PointF( (_physicalMin.X + offset.X), (_physicalMin.Y + offset.Y) );
 		}
